Stephan's review comments
