+++ tests/std/test.lst
@@ -262,7 +262,7 @@ tests\P0896R4_ranges_alg_generate_n
 tests\P0896R4_ranges_alg_heap
 tests\P0896R4_ranges_alg_is_permutation
 tests\P0896R4_ranges_alg_is_sorted
-tests\P0896R4_ranges_alg_lexi
+tests\P0896R4_ranges_alg_lexicographical_compare
 tests\P0896R4_ranges_alg_minmax
 tests\P0896R4_ranges_alg_mismatch
 tests\P0896R4_ranges_alg_move
